Fix fitZig input problems
The MRexperiments passed to fitZig did not have normalization factors, which are set when the function cumNorm is used on an MRexp. cumNorm seems to be applied to a base experiment and then then the count values are used to create new MRexps of various subdivision, but this does not bring the normFactors along. Normfactors are now set by calcNormFactors with the appropriate base table.
This change seemed to cause a change in the output object type, which required a different operator to access characteristics (@ instead of $)